requirements.txt
streamlit==1.32.0
tensorflow==2.12.0 
tensorflow-addons==0.19.0
numpy==1.23.5
Pillow==10.2.0
requests==2.31.0
torch
transformers==4.40.1
accelerate==0.27.2
sentencepiece==0.1.99
protobuf==3.20.3
